--- buildSrc/src/main/groovy/com/carrotsearch/gradle/randomizedtesting/SlowTestsConfiguration.groovy
@@ -1,14 +1,0 @@
-package com.carrotsearch.gradle.randomizedtesting
-
-class SlowTestsConfiguration {
-    int heartbeat = 0
-    int summarySize = 0
-
-    void heartbeat(int heartbeat) {
-        this.heartbeat = heartbeat
-    }
-
-    void summarySize(int summarySize) {
-        this.summarySize = summarySize
-    }
-}
