+++ core/build.gradle
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import com.carrotsearch.gradle.randomizedtesting.RandomizedTestingTask
+import com.carrotsearch.gradle.junit4.RandomizedTestingTask
 import org.elasticsearch.gradle.BuildPlugin
 import org.elasticsearch.gradle.test.RestSpecHack
